database fixes

+++ start.sh
@@ -1,11 +1,10 @@
 #!/usr/bin/env bash
 set -e
 
-if [ ! -f /var/data/fitsite.db ]; then
-if [ -f ./fitsite.db ]; then
-cp ./fitsite.db /var/data/fitsite.db
-fi
+mkdir -p /var/data
+if [ ! -f /var/data/fitsite.db ] && [ -f ./fitsite.db ]; then
+  cp ./fitsite.db /var/data/fitsite.db
 fi
 
-export ASPNETCORE_URLS=http://0.0.0.0:${PORT}
-exec dotnet ./publish/FitSite.dll
+export ASPNETCORE_URLS=http://0.0.0.0:${PORT:-10000}
+exec dotnet ./FitSite.dll
